XMLHttpRequest/open-url-multi-window-3.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: open() resolving URLs (multi-Window; 3; evil)</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      var test = async_test()
      function init() {
        test.step(function() {
          var client = new self[0].XMLHttpRequest()
          client.onreadystatechange = function() {
            test.step(function() {
              if(client.readyState == 4) {
                assert_equals(client.responseText, "bottom\n")
                test.done()
              }
            })
          }
          client.open("GET", "folder.txt")
          document.body.removeChild(document.getElementsByTagName("iframe")[0])
          client.send(null)
        })
      }
    </script>
    <iframe src="resources/init.htm"></iframe>
 </body>
</html>
